+++ src/main/java/org/opentripplanner/routing/algorithm/filterchain/ItineraryListFilterChainBuilder.java
@@ -7,10 +7,12 @@ import org.opentripplanner.routing.algorithm.filterchain.filter.AddMinSafeTransf
 import org.opentripplanner.routing.algorithm.filterchain.filter.DeletionFlaggingFilter;
 import org.opentripplanner.routing.algorithm.filterchain.filter.GroupByFilter;
 import org.opentripplanner.routing.algorithm.filterchain.filter.SortingFilter;
+import org.opentripplanner.routing.algorithm.filterchain.groupids.GroupByAllSameStations;
 import org.opentripplanner.routing.algorithm.filterchain.groupids.GroupByTripIdAndDistance;
 import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.LatestDepartureTimeFilter;
 import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.MaxLimitFilter;
 import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.NonTransitGeneralizedCostFilter;
+import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.OtherThanSameLegsMaxGeneralizedCostFilter;
 import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.RemoveBikerentalWithMostlyWalkingFilter;
 import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.RemoveParkAndRideWithMostlyWalkingFilter;
 import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.RemoveTransitIfStreetOnlyIsBetterFilter;
@@ -295,13 +297,35 @@ public class ItineraryListFilterChainBuilder {
             String name = "similar-legs-filter-" +
                     (int)(100d * it.groupByP) + "p-" + it.maxNumOfItinerariesPerGroup + "x";
 
+            List<ItineraryListFilter> nested = new ArrayList<>();
+
+            if (it.nestedGroupingByAllSameStations) {
+                final String innerGroupName = name + "-group-by-all-same-stations";
+                nested.add(new GroupByFilter<>(
+                    GroupByAllSameStations::new,
+                    List.of(
+                        new SortingFilter(new SortOnGeneralizedCost()),
+                        new DeletionFlaggingFilter( new MaxLimitFilter(innerGroupName, 1))
+                    )
+                ));
+            }
+
+            if (it.maxOtherLegsMultiplier > 1) {
+                nested.add(new DeletionFlaggingFilter(new OtherThanSameLegsMaxGeneralizedCostFilter(
+                    it.maxOtherLegsMultiplier
+                )));
+            }
+
+            nested.add(new SortingFilter(new SortOnGeneralizedCost()));
+            nested.add(new DeletionFlaggingFilter(new MaxLimitFilter(
+                name,
+                it.maxNumOfItinerariesPerGroup
+            )));
+
             groupByFilters.add(
                 new GroupByFilter<>(
                     itinerary -> new GroupByTripIdAndDistance(itinerary, it.groupByP),
-                    List.of(
-                        new SortingFilter(new SortOnGeneralizedCost()),
-                        new DeletionFlaggingFilter(new MaxLimitFilter(name, it.maxNumOfItinerariesPerGroup))
-                    )
+                    nested
                 )
             );
         }